+++ core/ledger/kvledger/txmgmt/validator/internal/tx_ops_preparation.go
@@ -12,7 +12,7 @@ import (
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/privacyenabledstate"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/rwsetutil"
 	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/statedb"
-	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/storageutil"
+	"github.com/hyperledger/fabric/core/ledger/kvledger/txmgmt/statemetadata"
 )
 
 func prepareTxOps(rwset *rwsetutil.TxRwSet, txht *version.Height,
@@ -105,7 +105,7 @@ func (txops txOps) applyMetadata(ns, coll string, metadataWrite *kvrwset.KVMetad
 	if metadataWrite.Entries == nil {
 		txops.metadataDelete(compositeKey{ns, coll, metadataWrite.Key})
 	} else {
-		metadataBytes, err := storageutil.SerializeMetadata(metadataWrite.Entries)
+		metadataBytes, err := statemetadata.Serialize(metadataWrite.Entries)
 		if err != nil {
 			return err
 		}
